# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Polideportivo API",
        "description": "Polideportivo API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/v1/deportes": {
            "get": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Get all the sports",
                "description": "Returns all sports as a JSON response",
                "operationId": "4c78774655c58577cc3c0790b58feb72",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Create a new sport",
                "description": "Create a new sport and return it as a JSON response",
                "operationId": "66f53c4c21ffb2f568dfc0df04e10c94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "hockey"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sport created successfully"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/deportes/{deporte}": {
            "get": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Get a specific sport",
                "description": "Returns a specific sport as a JSON response",
                "operationId": "7cd0744afebf445410c22c8a4206c440",
                "parameters": [
                    {
                        "name": "deporte",
                        "in": "path",
                        "description": "ID of the sport",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Update a specific sport",
                "description": "Update a specific sport and return it as a JSON         response",
                "operationId": "f401836f4ba6c3549c8c440f3b306ece",
                "parameters": [
                    {
                        "name": "deporte",
                        "in": "path",
                        "description": "ID of the sport",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "hockey"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sport updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Delete a specific sport",
                "description": "Delete a specific sport and return a JSON response",
                "operationId": "b0caf54b3180de072aa84e6cf6f1c357",
                "parameters": [
                    {
                        "name": "deporte",
                        "in": "path",
                        "description": "ID of the sport",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sport deleted successfully"
                    },
                    "404": {
                        "description": "Sport not found"
                    }
                }
            }
        },
        "/api/v1/pistas": {
            "get": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Get all the pistas",
                "description": "Returns all pistas as a JSON response",
                "operationId": "151736e5c2c8e07398d0fc0f52c95db6",
                "responses": {
                    "200": {
                        "description": "Request successful"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Create a new pista",
                "description": "Create a new pista and return it as a JSON response",
                "operationId": "8bfd2adafb4655c22de2b0be4ccb6bbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "deporte_id": {
                                        "type": "integer",
                                        "example": "6"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "example": "6"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pista created successfully"
                    }
                }
            }
        },
        "/api/v1/pistas/{pista}": {
            "get": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Get a specific pista",
                "description": "Returns a specific sport as a JSON response",
                "operationId": "81c176dc4724cebfdd5333072ceb154e",
                "parameters": [
                    {
                        "name": "pista",
                        "in": "path",
                        "description": "ID of the pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Update a specific pista",
                "description": "Update a specific pista and return it as a JSON         response",
                "operationId": "ec03709679cca502585bf0356bdb739c",
                "parameters": [
                    {
                        "name": "pista",
                        "in": "path",
                        "description": "ID of the pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "deporte_id": {
                                        "type": "integer",
                                        "example": "6"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "example": "6"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Delete a specific pista",
                "description": "Delete a specific pista and return a JSON response",
                "operationId": "38b8ea38077edfcf01dca737b6fc6602",
                "parameters": [
                    {
                        "name": "pista",
                        "in": "path",
                        "description": "ID of the pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pista deleted successfully"
                    },
                    "404": {
                        "description": "Pista not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "0": []
        }
    }
}